test(web): split query builder tests and parallelize client vitest
