+++ web/src/__tests__/server/queryBuilder-results-traces.servertest.ts
@@ -1,0 +1,1437 @@
+import {
+  QueryBuilder,
+  executeQuery,
+  createTrace,
+  createTracesCh,
+  randomUUID,
+  setupTracesWithObservations,
+} from "./queryBuilder.fixtures";
+import type { QueryType } from "./queryBuilder.fixtures";
+
+describe("queryBuilder", () => {
+  describe("query result correctness", () => {
+    describe("traces view", () => {
+      it("should count traces correctly", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          { name: "trace-1" },
+          { name: "trace-2" },
+          { name: "trace-3" },
+        ];
+
+        await setupTracesWithObservations(projectId, tracesData);
+
+        // Define query
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [],
+          metrics: [{ measure: "count", aggregation: "count" }],
+          filters: [],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(), // yesterday
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(), // tomorrow
+          orderBy: null,
+        };
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Assert
+        expect(result.data).toHaveLength(1);
+        expect(Number(result.data[0].count_count)).toBe(3);
+      });
+
+      it("should group traces by name and count correctly", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          { name: "chat-completion" },
+          { name: "chat-completion" },
+          { name: "embeddings" },
+          { name: "embeddings" },
+          { name: "embeddings" },
+        ];
+
+        await setupTracesWithObservations(projectId, tracesData);
+
+        // Define query
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }],
+          metrics: [{ measure: "count", aggregation: "count" }],
+          filters: [],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: null,
+        };
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Assert
+        expect(result.data).toHaveLength(2);
+
+        // Find and verify each name group
+        const chatCompletionRow = result.data.find(
+          (row: any) => row.name === "chat-completion",
+        );
+        expect(Number(chatCompletionRow.count_count)).toBe(2);
+
+        const embeddingsRow = result.data.find(
+          (row: any) => row.name === "embeddings",
+        );
+        expect(Number(embeddingsRow.count_count)).toBe(3);
+      });
+
+      it("should filter traces by name correctly", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          { name: "qa-bot" },
+          { name: "assistant" },
+          { name: "qa-bot" },
+          { name: "summarizer" },
+        ];
+
+        await setupTracesWithObservations(projectId, tracesData);
+
+        // Define query with a filter
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }],
+          metrics: [{ measure: "count", aggregation: "count" }],
+          filters: [
+            {
+              column: "name",
+              operator: "=",
+              value: "qa-bot",
+              type: "string",
+            },
+          ],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: null,
+        };
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Assert - should only return traces with name "qa-bot"
+        expect(result.data).toHaveLength(1);
+        expect(result.data[0].name).toBe("qa-bot");
+        expect(Number(result.data[0].count_count)).toBe(2);
+      });
+
+      it("should count observations per trace correctly", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          { name: "trace-with-many-obs", observationCount: 5 },
+          { name: "trace-with-few-obs", observationCount: 2 },
+          { name: "trace-with-no-obs", observationCount: 0 },
+        ];
+
+        await setupTracesWithObservations(projectId, tracesData);
+
+        // Define query to count observations
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }],
+          metrics: [{ measure: "observationsCount", aggregation: "sum" }],
+          filters: [],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: null,
+        };
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Assert
+        expect(result.data).toHaveLength(2);
+
+        // Find each trace by name and verify observation count
+        const manyObsTrace = result.data.find(
+          (row: any) => row.name === "trace-with-many-obs",
+        );
+        expect(Number(manyObsTrace.sum_observationsCount)).toBe(5);
+
+        const fewObsTrace = result.data.find(
+          (row: any) => row.name === "trace-with-few-obs",
+        );
+        expect(Number(fewObsTrace.sum_observationsCount)).toBe(2);
+      });
+
+      it("should use tags as dimension", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          { name: "trace-with-tag-a", tags: ["tag-a", "common-tag"] },
+          { name: "trace-with-tag-b", tags: ["tag-b", "common-tag"] },
+          { name: "trace-with-tag-c", tags: ["tag-c"] },
+          { name: "trace-with-no-tags", tags: [] },
+        ];
+
+        // Create traces with custom tags
+        const traces = [];
+        for (const data of tracesData) {
+          traces.push(
+            createTrace({
+              project_id: projectId,
+              name: data.name,
+              tags: data.tags,
+              timestamp: new Date().getTime(),
+            }),
+          );
+        }
+        await createTracesCh(traces);
+
+        // Define query with a filter for tags using "any of" operator
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "tags" }],
+          metrics: [{ measure: "count", aggregation: "count" }],
+          filters: [],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: null,
+        };
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        expect(result.data).toHaveLength(4);
+        // Expect one entry for all so index order does not matter
+        expect(Number(result.data[0].count_count)).toBe(1);
+      });
+
+      it("should filter traces by tags using 'any of' operator", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          { name: "trace-with-tag-a", tags: ["tag-a", "common-tag"] },
+          { name: "trace-with-tag-b", tags: ["tag-b", "common-tag"] },
+          { name: "trace-with-tag-c", tags: ["tag-c"] },
+          { name: "trace-with-no-tags", tags: [] },
+        ];
+
+        // Create traces with custom tags
+        const traces = [];
+        for (const data of tracesData) {
+          traces.push(
+            createTrace({
+              project_id: projectId,
+              name: data.name,
+              tags: data.tags,
+              timestamp: new Date().getTime(),
+            }),
+          );
+        }
+        await createTracesCh(traces);
+
+        // Define query with a filter for tags using "any of" operator
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }],
+          metrics: [{ measure: "count", aggregation: "count" }],
+          filters: [
+            {
+              column: "tags",
+              operator: "any of",
+              value: ["tag-a", "tag-b"],
+              type: "arrayOptions",
+            },
+          ],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: null,
+        };
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Assert - should only return traces with tag-a or tag-b
+        expect(result.data).toHaveLength(2);
+
+        // Find and verify each trace
+        const traceWithTagA = result.data.find(
+          (row: any) => row.name === "trace-with-tag-a",
+        );
+        expect(traceWithTagA).toBeDefined();
+        expect(Number(traceWithTagA.count_count)).toBe(1);
+
+        const traceWithTagB = result.data.find(
+          (row: any) => row.name === "trace-with-tag-b",
+        );
+        expect(traceWithTagB).toBeDefined();
+        expect(Number(traceWithTagB.count_count)).toBe(1);
+      });
+
+      it("should filter traces by tags using 'all of' operator", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          {
+            name: "trace-with-multiple-tags",
+            tags: ["tag-a", "tag-b", "common-tag"],
+          },
+          { name: "trace-with-tag-a-only", tags: ["tag-a", "common-tag"] },
+          { name: "trace-with-tag-b-only", tags: ["tag-b", "common-tag"] },
+          { name: "trace-with-other-tags", tags: ["tag-c", "common-tag"] },
+        ];
+
+        // Create traces with custom tags
+        const traces = [];
+        for (const data of tracesData) {
+          traces.push(
+            createTrace({
+              project_id: projectId,
+              name: data.name,
+              tags: data.tags,
+              timestamp: new Date().getTime(),
+            }),
+          );
+        }
+        await createTracesCh(traces);
+
+        // Define query with a filter for tags using "all of" operator
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }],
+          metrics: [{ measure: "count", aggregation: "count" }],
+          filters: [
+            {
+              column: "tags",
+              operator: "all of",
+              value: ["tag-a", "tag-b"],
+              type: "arrayOptions",
+            },
+          ],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: null,
+        };
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Assert - should only return traces with both tag-a and tag-b
+        expect(result.data).toHaveLength(1);
+
+        // Verify the trace has both tags
+        expect(result.data[0].name).toBe("trace-with-multiple-tags");
+        expect(Number(result.data[0].count_count)).toBe(1);
+      });
+
+      it("should filter traces by tags using 'none of' operator", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          { name: "trace-with-tag-a", tags: ["tag-a", "common-tag"] },
+          { name: "trace-with-tag-b", tags: ["tag-b", "common-tag"] },
+          {
+            name: "trace-with-other-tags",
+            tags: ["tag-c", "tag-d", "common-tag"],
+          },
+          { name: "trace-with-no-tags", tags: [] },
+        ];
+
+        // Create traces with custom tags
+        const traces = [];
+        for (const data of tracesData) {
+          traces.push(
+            createTrace({
+              project_id: projectId,
+              name: data.name,
+              tags: data.tags,
+              timestamp: new Date().getTime(),
+            }),
+          );
+        }
+        await createTracesCh(traces);
+
+        // Define query with a filter for tags using "none of" operator
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }],
+          metrics: [{ measure: "count", aggregation: "count" }],
+          filters: [
+            {
+              column: "tags",
+              operator: "none of",
+              value: ["tag-a", "tag-b"],
+              type: "arrayOptions",
+            },
+          ],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: null,
+        };
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Assert - should only return traces without tag-a or tag-b
+        expect(result.data).toHaveLength(2);
+
+        // Find and verify each trace
+        const traceWithOtherTags = result.data.find(
+          (row: any) => row.name === "trace-with-other-tags",
+        );
+        expect(traceWithOtherTags).toBeDefined();
+        expect(Number(traceWithOtherTags.count_count)).toBe(1);
+
+        const traceWithNoTags = result.data.find(
+          (row: any) => row.name === "trace-with-no-tags",
+        );
+        expect(traceWithNoTags).toBeDefined();
+        expect(Number(traceWithNoTags.count_count)).toBe(1);
+      });
+
+      it("should group by environment and calculate metrics correctly", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          {
+            name: "trace-prod-1",
+            environment: "production",
+            observationCount: 3,
+          },
+          {
+            name: "trace-prod-2",
+            environment: "production",
+            observationCount: 4,
+          },
+          {
+            name: "trace-dev-1",
+            environment: "development",
+            observationCount: 2,
+          },
+          {
+            name: "trace-dev-2",
+            environment: "development",
+            observationCount: 1,
+          },
+        ];
+
+        await setupTracesWithObservations(projectId, tracesData);
+
+        // Define query to group by environment
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "environment" }],
+          metrics: [
+            { measure: "count", aggregation: "count" },
+            { measure: "observationsCount", aggregation: "sum" },
+          ],
+          filters: [],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: null,
+        };
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Assert
+        expect(result.data).toHaveLength(2);
+
+        // Verify production environment data
+        const prodEnv = result.data.find(
+          (row: any) => row.environment === "production",
+        );
+        expect(Number(prodEnv.count_count)).toBe(2); // 2 traces
+        expect(Number(prodEnv.sum_observationsCount)).toBe(7); // 3+4 observations
+
+        // Verify development environment data
+        const devEnv = result.data.find(
+          (row: any) => row.environment === "development",
+        );
+        expect(Number(devEnv.count_count)).toBe(2); // 2 traces
+        expect(Number(devEnv.sum_observationsCount)).toBe(3); // 2+1 observations
+      });
+
+      it("should handle multiple dimensions (name and environment) correctly", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          { name: "chat", environment: "production", observationCount: 2 },
+          { name: "chat", environment: "development", observationCount: 1 },
+          {
+            name: "embeddings",
+            environment: "production",
+            observationCount: 3,
+          },
+          {
+            name: "embeddings",
+            environment: "development",
+            observationCount: 2,
+          },
+        ];
+
+        await setupTracesWithObservations(projectId, tracesData);
+
+        // Define query with multiple dimensions
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }, { field: "environment" }],
+          metrics: [{ measure: "count", aggregation: "count" }],
+          filters: [],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: null,
+        };
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Assert
+        expect(result.data).toHaveLength(4); // 2 names × 2 environments = 4 combinations
+
+        // Verify each combination
+        const chatProd = result.data.find(
+          (row: any) => row.name === "chat" && row.environment === "production",
+        );
+        expect(Number(chatProd.count_count)).toBe(1);
+
+        const chatDev = result.data.find(
+          (row: any) =>
+            row.name === "chat" && row.environment === "development",
+        );
+        expect(Number(chatDev.count_count)).toBe(1);
+
+        const embeddingsProd = result.data.find(
+          (row: any) =>
+            row.name === "embeddings" && row.environment === "production",
+        );
+        expect(Number(embeddingsProd.count_count)).toBe(1);
+
+        const embeddingsDev = result.data.find(
+          (row: any) =>
+            row.name === "embeddings" && row.environment === "development",
+        );
+        expect(Number(embeddingsDev.count_count)).toBe(1);
+      });
+
+      it("should handle multiple metrics correctly", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          { name: "multi-metric-test", observationCount: 10 },
+          { name: "multi-metric-test", observationCount: 20 },
+        ];
+
+        await setupTracesWithObservations(projectId, tracesData);
+
+        // Define query with multiple metrics
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }],
+          metrics: [
+            { measure: "count", aggregation: "count" },
+            { measure: "observationsCount", aggregation: "sum" },
+            { measure: "observationsCount", aggregation: "avg" },
+            { measure: "observationsCount", aggregation: "max" },
+            { measure: "observationsCount", aggregation: "min" },
+          ],
+          filters: [],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: null,
+        };
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Assert
+        expect(result.data).toHaveLength(1);
+
+        const row = result.data[0];
+        expect(row.name).toBe("multi-metric-test");
+        expect(Number(row.count_count)).toBe(2); // 2 traces
+        expect(Number(row.sum_observationsCount)).toBe(30); // 10+20 observations
+        expect(row.avg_observationsCount).toBe(15); // (10+20)/2 average
+        expect(Number(row.max_observationsCount)).toBe(20); // max is 20
+        expect(Number(row.min_observationsCount)).toBe(10); // min is 10
+      });
+
+      it("should order by a dimension field correctly", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          { name: "trace-c", environment: "staging" },
+          { name: "trace-a", environment: "production" },
+          { name: "trace-b", environment: "development" },
+        ];
+
+        await setupTracesWithObservations(projectId, tracesData);
+
+        // Define query with orderBy on a dimension
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }, { field: "environment" }],
+          metrics: [{ measure: "count", aggregation: "count" }],
+          filters: [],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: [{ field: "name", direction: "asc" }],
+        };
+
+        // Execute query
+        const queryBuilder = new QueryBuilder();
+        const { query: compiledQuery } = await queryBuilder.build(
+          query,
+          projectId,
+        );
+
+        // Verify ORDER BY clause is present in the query
+        expect(compiledQuery).toContain("ORDER BY name asc");
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Assert - results should be ordered by name alphabetically
+        expect(result.data).toHaveLength(3);
+        expect(result.data[0].name).toBe("trace-a");
+        expect(result.data[1].name).toBe("trace-b");
+        expect(result.data[2].name).toBe("trace-c");
+      });
+
+      it("should order by a metric field correctly", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          { name: "trace-low", observationCount: 2 },
+          { name: "trace-high", observationCount: 10 },
+          { name: "trace-medium", observationCount: 5 },
+        ];
+
+        await setupTracesWithObservations(projectId, tracesData);
+
+        // Define query with orderBy on a metric
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }],
+          metrics: [{ measure: "observationsCount", aggregation: "sum" }],
+          filters: [],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: [{ field: "sum_observationsCount", direction: "desc" }],
+        };
+
+        // Execute query
+        const queryBuilder = new QueryBuilder();
+        const { query: compiledQuery } = await queryBuilder.build(
+          query,
+          projectId,
+        );
+
+        // Verify ORDER BY clause is present in the query
+        expect(compiledQuery).toContain("ORDER BY sum_observationsCount desc");
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Assert - results should be ordered by observation count descending
+        expect(result.data).toHaveLength(3);
+        expect(result.data[0].name).toBe("trace-high"); // 10 observations
+        expect(result.data[1].name).toBe("trace-medium"); // 5 observations
+        expect(result.data[2].name).toBe("trace-low"); // 2 observations
+      });
+
+      it("should order by multiple fields correctly", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          { name: "trace-a", environment: "production", observationCount: 5 },
+          { name: "trace-b", environment: "production", observationCount: 2 },
+          { name: "trace-c", environment: "development", observationCount: 7 },
+          { name: "trace-d", environment: "development", observationCount: 3 },
+        ];
+
+        await setupTracesWithObservations(projectId, tracesData);
+
+        // Define query with multiple orderBy fields
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "environment" }, { field: "name" }],
+          metrics: [{ measure: "observationsCount", aggregation: "sum" }],
+          filters: [],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: [
+            { field: "environment", direction: "asc" },
+            { field: "sum_observationsCount", direction: "desc" },
+          ],
+        };
+
+        const queryBuilder = new QueryBuilder();
+        const { query: compiledQuery } = await queryBuilder.build(
+          query,
+          projectId,
+        );
+
+        // Verify ORDER BY clause is present in the query with both fields
+        expect(compiledQuery).toContain(
+          "ORDER BY environment asc, sum_observationsCount desc",
+        );
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+        // Assert - results should be ordered by environment (asc) and then by observation count (desc)
+        expect(result.data).toHaveLength(4);
+
+        // Development environment should come first (alphabetically)
+        expect(result.data[0].environment).toBe("development");
+        expect(result.data[1].environment).toBe("development");
+        // Within development, higher observation count first
+        expect(result.data[0].name).toBe("trace-c"); // 7 observations
+        expect(result.data[1].name).toBe("trace-d"); // 3 observations
+
+        // Production environment should come second
+        expect(result.data[2].environment).toBe("production");
+        expect(result.data[3].environment).toBe("production");
+        // Within production, higher observation count first
+        expect(result.data[2].name).toBe("trace-a"); // 5 observations
+        expect(result.data[3].name).toBe("trace-b"); // 2 observations
+      });
+
+      it("should handle default ordering when no orderBy is specified", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const now = new Date();
+        const yesterday = new Date(now);
+        yesterday.setDate(yesterday.getDate() - 1);
+
+        // Create traces with time dimension
+        const traces = [
+          createTrace({
+            project_id: projectId,
+            name: "trace-day2",
+            environment: "default",
+            timestamp: now.getTime(),
+          }),
+          createTrace({
+            project_id: projectId,
+            name: "trace-day1",
+            environment: "default",
+            timestamp: yesterday.getTime(),
+          }),
+        ];
+
+        await createTracesCh(traces);
+
+        // Define query with time dimension but no explicit orderBy
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }],
+          metrics: [{ measure: "count", aggregation: "count" }],
+          filters: [],
+          timeDimension: {
+            granularity: "day",
+          },
+          fromTimestamp: yesterday.toISOString(),
+          toTimestamp: new Date(now.getTime() + 86400000).toISOString(),
+          orderBy: null,
+        };
+
+        const queryBuilder = new QueryBuilder();
+        const { query: compiledQuery } = await queryBuilder.build(
+          query,
+          projectId,
+        );
+
+        // Verify ORDER BY clause includes default time dimension ordering
+        expect(compiledQuery).toContain("ORDER BY time_dimension asc");
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Results should be ordered by time dimension (ascending)
+        expect(result.data.length).toBeGreaterThan(0);
+
+        // Convert time_dimension strings to Date objects for easier testing
+        const dates = result.data.map(
+          (row: any) => new Date(row.time_dimension),
+        );
+
+        // Check that dates are in ascending order
+        for (let i = 1; i < dates.length; i++) {
+          expect(dates[i].getTime()).toBeGreaterThanOrEqual(
+            dates[i - 1].getTime(),
+          );
+        }
+      });
+
+      it("should use first metric for default ordering when no time dimension", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          { name: "trace-low", observationCount: 2 },
+          { name: "trace-high", observationCount: 10 },
+          { name: "trace-medium", observationCount: 5 },
+        ];
+
+        await setupTracesWithObservations(projectId, tracesData);
+
+        // Define query with metrics but no explicit orderBy and no time dimension
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }],
+          metrics: [{ measure: "observationsCount", aggregation: "sum" }],
+          filters: [],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: null,
+        };
+
+        const queryBuilder = new QueryBuilder();
+        const { query: compiledQuery } = await queryBuilder.build(
+          query,
+          projectId,
+        );
+
+        // Verify ORDER BY clause includes default metric ordering (descending)
+        expect(compiledQuery).toContain("ORDER BY sum_observationsCount desc");
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Results should be ordered by observation count descending (default for metrics)
+        expect(result.data).toHaveLength(3);
+        expect(result.data[0].name).toBe("trace-high"); // 10 observations
+        expect(result.data[1].name).toBe("trace-medium"); // 5 observations
+        expect(result.data[2].name).toBe("trace-low"); // 2 observations
+      });
+
+      it("should use first dimension for default ordering when no metrics and no time dimension", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          { name: "trace-c", environment: "production" },
+          { name: "trace-a", environment: "production" },
+          { name: "trace-b", environment: "production" },
+        ];
+
+        await setupTracesWithObservations(projectId, tracesData);
+
+        // Define query with dimensions but no metrics, no time dimension, and no explicit orderBy
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }],
+          metrics: [],
+          filters: [],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: null,
+        };
+
+        const queryBuilder = new QueryBuilder();
+        const { query: compiledQuery } = await queryBuilder.build(
+          query,
+          projectId,
+        );
+
+        // Verify ORDER BY clause includes default dimension ordering (ascending)
+        expect(compiledQuery).toContain("ORDER BY name asc");
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Results should be ordered by name ascending (default for dimensions)
+        expect(result.data).toHaveLength(3);
+        expect(result.data[0].name).toBe("trace-a");
+        expect(result.data[1].name).toBe("trace-b");
+        expect(result.data[2].name).toBe("trace-c");
+      });
+
+      it("should filter with the LIKE operator correctly", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          { name: "prefix-value-1" },
+          { name: "prefix-value-2" },
+          { name: "different-name" },
+        ];
+
+        await setupTracesWithObservations(projectId, tracesData);
+
+        // Define query with LIKE filter
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }],
+          metrics: [{ measure: "count", aggregation: "count" }],
+          filters: [
+            {
+              column: "name",
+              operator: "starts with",
+              value: "prefix-",
+              type: "string",
+            },
+          ],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: null,
+        };
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Assert
+        expect(result.data).toHaveLength(2);
+        expect(
+          result.data.every((row: any) => row.name.startsWith("prefix-")),
+        ).toBe(true);
+        expect(result.data.map((row: any) => row.name).sort()).toEqual([
+          "prefix-value-1",
+          "prefix-value-2",
+        ]);
+      });
+
+      it("should handle queries with zero dimensions and zero metrics", async () => {
+        // Setup - just create a few traces
+        const projectId = randomUUID();
+        const tracesData = [{ name: "trace-1" }, { name: "trace-2" }];
+
+        await setupTracesWithObservations(projectId, tracesData);
+
+        // Define query with no dimensions and no metrics
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [],
+          metrics: [],
+          filters: [],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: null,
+        };
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Assert
+        expect(result.data).toHaveLength(1);
+        expect(Number(result.data[0].count)).toBe(2); // default count metric should be used
+      });
+
+      it("should fill gaps in time series data with WITH FILL", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const now = new Date();
+
+        // Create traces at specific dates with intentional gaps
+        const day1 = new Date(now);
+        day1.setDate(day1.getDate() - 3);
+        day1.setHours(12, 0, 0, 0);
+
+        const day3 = new Date(now);
+        day3.setDate(day3.getDate() - 1);
+        day3.setHours(12, 0, 0, 0);
+
+        // Skip day2 to create a gap
+
+        const traces = [
+          // Day 1 trace
+          createTrace({
+            project_id: projectId,
+            name: "trace-test",
+            environment: "default",
+            timestamp: day1.getTime(),
+          }),
+          // Day 3 trace
+          createTrace({
+            project_id: projectId,
+            name: "trace-test",
+            environment: "default",
+            timestamp: day3.getTime(),
+          }),
+        ];
+
+        await createTracesCh(traces);
+
+        // Set from timestamp to day before day1 and to timestamp to day after day3
+        const fromDate = new Date(day1);
+        fromDate.setDate(fromDate.getDate() - 1);
+
+        const toDate = new Date(day3);
+        toDate.setDate(toDate.getDate() + 1);
+
+        // Define query with time dimension and daily granularity
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }],
+          metrics: [{ measure: "count", aggregation: "count" }],
+          filters: [
+            {
+              column: "name",
+              operator: "=",
+              value: "trace-test",
+              type: "string",
+            },
+          ],
+          timeDimension: {
+            granularity: "day", // Use day granularity for clear gap visibility
+          },
+          fromTimestamp: fromDate.toISOString(),
+          toTimestamp: toDate.toISOString(),
+          orderBy: null,
+        };
+
+        const queryBuilder = new QueryBuilder();
+        const { query: compiledQuery } = await queryBuilder.build(
+          query,
+          projectId,
+        );
+
+        // Verify WITH FILL clause is present in the query
+        expect(compiledQuery).toContain("WITH FILL");
+        expect(compiledQuery).toContain("STEP INTERVAL 1 DAY");
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Expected to have 3 days in the result (including the filled gap)
+        expect(result.data.length).toBeGreaterThanOrEqual(3);
+
+        // Convert time_dimension strings to Date objects for easier testing
+        const resultDates = result.data.map((row: any) => {
+          return {
+            date: new Date(row.time_dimension),
+            count: row.count_count ? parseInt(row.count_count) : 0,
+          };
+        });
+
+        // Find day1 and day3 (where we have actual traces)
+        const day1Row = resultDates.find(
+          (row: any) =>
+            row.date.getDate() === day1.getDate() &&
+            row.date.getMonth() === day1.getMonth(),
+        );
+
+        const day3Row = resultDates.find(
+          (row: any) =>
+            row.date.getDate() === day3.getDate() &&
+            row.date.getMonth() === day3.getMonth(),
+        );
+
+        // Find day2 (gap day that should be filled with zeros)
+        const day2 = new Date(day1);
+        day2.setDate(day2.getDate() + 1);
+
+        const day2Row = resultDates.find(
+          (row: any) =>
+            row.date.getDate() === day2.getDate() &&
+            row.date.getMonth() === day2.getMonth(),
+        );
+
+        // Assert actual data and filled gaps
+        expect(day1Row?.count).toBe(1); // Actual trace for day 1
+        expect(day3Row?.count).toBe(1); // Actual trace for day 3
+        expect(day2Row).toBeDefined(); // Day 2 should exist (filled)
+        expect(day2Row?.count).toBe(0); // Day 2 should be filled with 0
+
+        // Check that all dates form a continuous sequence
+        for (let i = 1; i < resultDates.length; i++) {
+          const prevDate = resultDates[i - 1].date;
+          const currDate = resultDates[i].date;
+
+          // Difference should be approximately 1 day (86400000 ms)
+          const diffMs = currDate.getTime() - prevDate.getTime();
+          expect(diffMs).toBeGreaterThanOrEqual(86000000); // ~1 day with some tolerance
+          expect(diffMs).toBeLessThanOrEqual(86800000); // ~1 day with some tolerance
+        }
+      });
+
+      it("should group traces by name and time dimension correctly", async () => {
+        // Setup
+        const projectId = randomUUID();
+
+        // Create traces with specific timestamps and names
+        const now = new Date();
+        const yesterday = new Date(now);
+        yesterday.setDate(yesterday.getDate() - 1);
+        const dayBeforeYesterday = new Date(now);
+        dayBeforeYesterday.setDate(dayBeforeYesterday.getDate() - 2);
+
+        const traces = [];
+        // Today's traces - two different names
+        traces.push(
+          createTrace({
+            project_id: projectId,
+            name: "chat-completion",
+            environment: "default",
+            timestamp: now.getTime(),
+          }),
+        );
+        traces.push(
+          createTrace({
+            project_id: projectId,
+            name: "chat-completion",
+            environment: "default",
+            timestamp: now.getTime(),
+          }),
+        );
+        traces.push(
+          createTrace({
+            project_id: projectId,
+            name: "embeddings",
+            environment: "default",
+            timestamp: now.getTime(),
+          }),
+        );
+
+        // Yesterday's traces
+        traces.push(
+          createTrace({
+            project_id: projectId,
+            name: "chat-completion",
+            environment: "default",
+            timestamp: yesterday.getTime(),
+          }),
+        );
+        traces.push(
+          createTrace({
+            project_id: projectId,
+            name: "embeddings",
+            environment: "default",
+            timestamp: yesterday.getTime(),
+          }),
+        );
+
+        // Day before yesterday's traces
+        traces.push(
+          createTrace({
+            project_id: projectId,
+            name: "embeddings",
+            environment: "default",
+            timestamp: dayBeforeYesterday.getTime(),
+          }),
+        );
+
+        await createTracesCh(traces);
+
+        // Define query with time dimension and name dimension
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }],
+          metrics: [{ measure: "count", aggregation: "count" }],
+          filters: [],
+          timeDimension: {
+            granularity: "day", // Group by day
+          },
+          fromTimestamp: dayBeforeYesterday.toISOString(),
+          toTimestamp: new Date(now.getTime() + 86400000).toISOString(), // Include tomorrow
+          orderBy: null,
+        };
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Assert - should have 5 combinations (chat-today, chat-yesterday, embeddings-today, embeddings-yesterday, embeddings-dayBefore)
+        expect(result.data).toHaveLength(5);
+
+        // Check chat-completion counts by day
+        const chatCompletionToday = result.data.find((row: any) => {
+          const rowDate = new Date(row.time_dimension);
+          const today = new Date(now);
+          return (
+            row.name === "chat-completion" &&
+            rowDate.getDate() === today.getDate() &&
+            rowDate.getMonth() === today.getMonth() &&
+            rowDate.getFullYear() === today.getFullYear()
+          );
+        });
+        expect(Number(chatCompletionToday?.count_count)).toBe(2); // 2 chat-completion traces today
+
+        const chatCompletionYesterday = result.data.find((row: any) => {
+          const rowDate = new Date(row.time_dimension);
+          const yesterday = new Date(now);
+          yesterday.setDate(yesterday.getDate() - 1);
+          return (
+            row.name === "chat-completion" &&
+            rowDate.getDate() === yesterday.getDate() &&
+            rowDate.getMonth() === yesterday.getMonth() &&
+            rowDate.getFullYear() === yesterday.getFullYear()
+          );
+        });
+        expect(Number(chatCompletionYesterday?.count_count)).toBe(1); // 1 chat-completion trace yesterday
+
+        // Check embeddings counts by day
+        const embeddingsToday = result.data.find((row: any) => {
+          const rowDate = new Date(row.time_dimension);
+          const today = new Date(now);
+          return (
+            row.name === "embeddings" &&
+            rowDate.getDate() === today.getDate() &&
+            rowDate.getMonth() === today.getMonth() &&
+            rowDate.getFullYear() === today.getFullYear()
+          );
+        });
+        expect(Number(embeddingsToday?.count_count)).toBe(1); // 1 embeddings trace today
+
+        const embeddingsYesterday = result.data.find((row: any) => {
+          const rowDate = new Date(row.time_dimension);
+          const yesterday = new Date(now);
+          yesterday.setDate(yesterday.getDate() - 1);
+          return (
+            row.name === "embeddings" &&
+            rowDate.getDate() === yesterday.getDate() &&
+            rowDate.getMonth() === yesterday.getMonth() &&
+            rowDate.getFullYear() === yesterday.getFullYear()
+          );
+        });
+        expect(Number(embeddingsYesterday?.count_count)).toBe(1); // 1 embeddings trace yesterday
+
+        const embeddingsDayBefore = result.data.find((row: any) => {
+          const rowDate = new Date(row.time_dimension);
+          const dayBefore = new Date(now);
+          dayBefore.setDate(dayBefore.getDate() - 2);
+          return (
+            row.name === "embeddings" &&
+            rowDate.getDate() === dayBefore.getDate() &&
+            rowDate.getMonth() === dayBefore.getMonth() &&
+            rowDate.getFullYear() === dayBefore.getFullYear()
+          );
+        });
+        expect(Number(embeddingsDayBefore?.count_count)).toBe(1); // 1 embeddings trace day before yesterday
+      });
+
+      it("should ensure time_dimension adheres to ISO8601 date format", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const now = new Date();
+
+        // Create traces with specific timestamps
+        const traces = [
+          createTrace({
+            project_id: projectId,
+            name: "test-trace-1",
+            environment: "production",
+            timestamp: now.getTime(),
+          }),
+          createTrace({
+            project_id: projectId,
+            name: "test-trace-2",
+            environment: "production",
+            timestamp: now.getTime() - 86400000, // 1 day ago
+          }),
+        ];
+        await createTracesCh(traces);
+
+        // Define query with time dimension
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }],
+          metrics: [{ measure: "count", aggregation: "count" }],
+          filters: [],
+          timeDimension: {
+            granularity: "hour",
+          },
+          fromTimestamp: new Date(now.getTime() - 172800000).toISOString(), // 2 days ago
+          toTimestamp: now.toISOString(),
+          orderBy: null,
+        };
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        // Verify we have results
+        expect(result.data.length).toBeGreaterThan(0);
+
+        // Check that time_dimension values adhere to ISO8601 format
+        for (const row of result.data) {
+          expect(row).toHaveProperty("time_dimension");
+
+          // ISO8601 regex pattern
+          // This pattern matches ISO8601 dates in format: YYYY-MM-DD or YYYY-MM-DDThh:mm:ss.sssZ
+          const iso8601Pattern =
+            /^\d{4}-\d{2}-\d{2}(T\d{2}:\d{2}:\d{2}(\.\d+)?(Z|[+-]\d{2}:\d{2})?)?$/;
+
+          expect(row.time_dimension).toMatch(iso8601Pattern);
+
+          // Verify that JavaScript Date constructor can parse the time_dimension
+          const date = new Date(row.time_dimension);
+          expect(date).toBeInstanceOf(Date);
+          expect(date.toString()).not.toBe("Invalid Date");
+        }
+      });
+
+      it("should filter traces by metadata correctly", async () => {
+        // Setup
+        const projectId = randomUUID();
+        const tracesData = [
+          {
+            name: "trace-with-metadata-1",
+            metadata: { customer: "test1" },
+          },
+          {
+            name: "trace-with-metadata-2",
+            metadata: { customer: "test2" },
+          },
+          {
+            name: "trace-without-metadata",
+            metadata: undefined,
+          },
+        ];
+
+        // Create traces with metadata
+        const traces = [];
+        for (const data of tracesData) {
+          const trace = await createTrace({
+            id: randomUUID(),
+            name: data.name,
+            project_id: projectId,
+            metadata: data.metadata,
+          });
+          traces.push(trace);
+        }
+
+        await createTracesCh(traces);
+
+        // Define query with metadata filter
+        const query: QueryType = {
+          view: "traces",
+          dimensions: [{ field: "name" }],
+          metrics: [{ measure: "count", aggregation: "count" }],
+          filters: [
+            {
+              column: "metadata",
+              operator: "contains",
+              key: "customer",
+              value: "test",
+              type: "stringObject",
+            },
+          ],
+          timeDimension: null,
+          fromTimestamp: new Date(
+            new Date().setDate(new Date().getDate() - 1),
+          ).toISOString(),
+          toTimestamp: new Date(
+            new Date().setDate(new Date().getDate() + 1),
+          ).toISOString(),
+          orderBy: null,
+        };
+
+        // Execute query
+        const result: { data: Array<any> } = { data: [] };
+        result.data = await executeQuery(projectId, query);
+
+        expect(result.data).toHaveLength(2);
+        expect(result.data[0].name).toBe("trace-with-metadata-1");
+        expect(Number(result.data[0].count_count)).toBe(1);
+      });
+    });
+  });
+});
